Create gates in bulk.

Creates multiple gates in a single API call. This is faster than calling createXxxGate in a loop.

createGates(experimentId, gates)

Arguments

experimentId
The ID of the experiment to which to add the gate.

gates
list() of gates.

Details

Each gate must be specified as a list() object conforming to our API documentation; however, some properties will have defaults automatically set if not specified (model.locked, model.label, gid, parentPopulationId, tailoredPerFile, fcsFileId, experimentId).

Up to ~1,000 gates may be created at once.

Unlike createXxxGate, this function cannot automatically create populations corresponding to each gate. This function is primarily suited to tailoring a large number of gates.

Examples

# NOT RUN {
g1 = list(type = "RectangleGate", xChannel = "FSC-A", yChannel = "SSC-A",
  model = list(rectangle = list(x1 = 1, x2 = 100, y1 = 1, y2 = 100)))
g2 = list(type = "PolygonGate", xChannel = "FSC-A", yChannel = "SSC-A",
  model = list(polygon = list(vertices = c(c(1, 2), c(30, 40), c(50, 60)))))
g3 = list(type = "RangeGate", xChannel = "V450-480-A",
  model = list(range = list(x1 = 1, x2 = 100, y = 0.5)))

createGates(experimentId, c(g1, g2, g3))
# }