Create population

Creates a population.

createPopulation(experimentId, name, gates, terminalGateGid, parentId = NULL)

Arguments

experimentId

The ID of the experiment to which to add the population.

name

The name of the population.

gates

The list of gates comprising the population. This should be in the form list(`$and` = c("gid1", "gid2")), where "gid1" and "gid2" are the Group IDs (gids) of the gates. Any combination of $and, $or, $not and $xor can be used; however, only populations with the form list(`$and` = c(...gates)) can be displayed in the Web UI at this time. Alternatively, this may be a JSON string (may be easier than constructing the correct R expression).

terminalGateGid

The Group ID (gid) of the gate that differentiates this population from its parent. Because complex populations that deviate from the list(`$and` = c(...gates)) form cannot currently be displayed in the Web UI, those populations may set this to NULL. In the future, this will be required to be a list of GIDs for those populations.

parentId

ID of the parent population. Use NULL for the "ungated" population.

Examples

if (FALSE) {
gid1 <- "59262d84b1a1fc1193f12b0e"
createPopulation(experimentId, "Singlets", list(`$and` = c(gid1)), gid1)
}