Create quadrant gate

Creates a quadrant gate. Quadrant gates have four sectors (upper-right, upper-left, lower-left, lower-right), each with a unique gid and name.

createQuadrantGate(experimentId, xChannel, yChannel, name, x, y,
  labels = NULL, gid = generateId(), gids = replicate(4,
  generateId()), parentPopulationId = NULL, parentPopulation = NULL,
  tailoredPerFile = FALSE, fcsFileId = NULL, fcsFile = NULL,
  locked = FALSE, createPopulations = TRUE)

Arguments

experimentId
The ID of the experiment to which to add the gate.

xChannel
The name of the x channel to which the gate applies.

yChannel
The name of the y channel to which the gate applies.

name
The name of the gate. Each sector is named with this parameter and a quadrant flag (UR, UL, LL, LR).

x
The x coordinate (after the channel's scale has been applied).

y
The y coordinate (after the channel's scale has been applied).

labels
Positions of the quadrant labels. A list of four length-2 vectors in the order: UR, UL, LL, LR. These are set automatically to the plot corners.

gid
Top-level group ID of the gate, used for tailoring. If this is not specified, then a new Group ID will be created. For compound gates, "gid" refers to the top-level GID. Each sector has a unique model gid and name to which populations must refer.

gids
Group IDs of each sector, assigned to model.gids.

parentPopulationId
ID of the parent population. Use NULL for the "ungated" population. If specified, do not specify parentPopulation.

parentPopulation
Name of the parent population. An attempt will be made to find the population by name. If zero or more than one population exists with the name, an error will be thrown. If specified, do not specify parentPopulationId.

tailoredPerFile
Whether this gate is tailored per FCS file.

fcsFileId
ID of FCS file, if tailored per file. Use NULL for the global gate in a tailored gate group. If specified, do not specify fcsFile.

fcsFile
Name of FCS file, if tailored per file. An attempt will be made to find the file by name. If zero or more than one file exists with the name, an error will be thrown. If specified, do not specify fcsFileId.

locked
Prevents modification of the gate via the web interface.

createPopulations
Automatically create corresponding population.

angles
Angles at which the quadrant lines appear, in radians. Zero (0) points horizontally to the right; angles proceed counter-clockwise. Currently these must be 0, pi / 2, pi and 3 * pi / 2.

Examples

# NOT RUN {
createQuadrantGate(experimentId, "FSC-A", "FSC-W", "my gate", 160000, 200000)
# }