Convert a flowCore object to a CellEngine object

Creates a CellEngine object from a flowCore object, saving it to CellEngine.

fromFlowCore(flowCoreObj, experimentId, name = NULL, ...)

Arguments

flowCoreObj

A flowCore object.

experimentId

The ID of the experiment in which to create the object.

name

The name to use for the CellEngine object. If not set, it will default to the identifier of the flowCore object.

...

Other parameters passed to createRectangleGate, createPolygonGate, etc.

Examples

if (FALSE) {
# Converting a flowCore rectangleGate:
rectGate <- rectangleGate(
  filterId = "Gate 1",
  "FL1-H" = c(0, 12), "FL2-H" = c(0, 12)
)
experimentId <- "5d2f8b4b21fd0676fb3a6a8c"
fromFlowCore(rectGate, experimentId, name = "my gate")
# "my gate" should now exist in the CellEngine experiment.
}