Create polygon gate

Creates a polygon gate.

createPolygonGate(
  experimentId,
  xChannel,
  yChannel,
  name,
  xVertices = c(),
  yVertices = c(),
  label = c(mean(xVertices), mean(yVertices)),
  gid = generateId(),
  parentPopulationId = NULL,
  parentPopulation = NULL,
  tailoredPerFile = FALSE,
  fcsFileId = NULL,
  fcsFile = NULL,
  locked = FALSE,
  createPopulation = TRUE
)

Arguments

experimentId
The ID of the experiment to which to add the gate.

xChannel
The name of the x channel to which the gate applies.

yChannel
The name of the y channel to which the gate applies.

name
The name of the gate.

xVertices
List of x coordinates for the polygon's vertices.

yVertices
List of y coordinates for the polygon's vertices.

label
Position of the label. Defaults to the midpoint of the gate.

gid
Group ID of the gate, used for tailoring. If this is not specified, then a new Group ID will be created.

parentPopulationId
ID of the parent population. Use NULL for the "ungated" population. If specified, do not specify parentPopulation.

parentPopulation
Name of the parent population. An attempt will be made to find the population by name. If zero or more than one population exists with the name, an error will be thrown. If specified, do not specify parentPopulationId.

tailoredPerFile
Whether or not this gate is tailored per FCS file.

fcsFileId
ID of FCS file, if tailored per file. Use NULL for the global gate in a tailored gate group. If specified, do not specify fcsFile.

fcsFile
Name of FCS file, if tailored per file. An attempt will be made to find the file by name. If zero or more than one file exists with the name, an error will be thrown. If specified, do not specify fcsFileId.

locked
Prevents modification of the gate via the web interface.

createPopulation
Automatically create corresponding population.

Examples

if (FALSE) {
createPolygonGate(experimentId, "FSC-A", "FSC-W", "my gate", c(1, 2, 3), c(4, 5, 6))
}