Get FCS File Events

Retrieves events from an FCS file or gated population as either an FCS or TSV file.

getEvents(experimentId, fcsFileId, populationId = NULL,
  compensation = NULL, compensatedQ = FALSE, headerQ = FALSE,
  format = "FCS", destination = NULL, overwrite = FALSE,
  subsampling = list())

Arguments

experimentId
ID of experiment.

fcsFileId
ID of FCS file.

populationId
Optional ID of population, for gated events.

compensation
ID of compensation or special compensation type (UNCOMPENSATED or FILE_INTERNAL) to use for gating. This should generally match what you used to create your gates. Not required if compensatedQ is FALSE and exporting ungated files.

compensatedQ
If TRUE, applies the compensation specified in compensation to the exported events. For TSV format, the numerical values will be the compensated values. For FCS format, the numerical values will be unchanged, but the file header will contain the compensation as the spill string (file-internal compensation).

headerQ
for TSV format only: if true, file will contain a header row.

format
One of "TSV" or "FCS".

destination
Optional, if specified, write the file to the specified path instead of returning it as a binary blob (FCS) or data.frame (TSV).

overwrite
Optional, if a destination is specified, allows destination file to be overwritten.

subsampling
Optional, subsampling parameters in the form list(preSubsampleN = number, preSubsampleP = number, postSubsampleN = number, postSubsampleP = number, seed = number). Subsample-N options specify absolute subsampling values, and subsample-P options specify a fractional subsampling value (0 to 1); specify only one pre- and/or one post-subsample option. Specify a seed for reproducible downsampling.

Details

Tip: FCS files are more compact than TSV files. Use that format for faster downloads. Use a library such as flowCore to parse FCS files.

Examples

# NOT RUN {
# Returns the FCS file as a binary blob (requires parsing before use):
getEvents(experimentId, fcsFileId)

# Returns a data.frame:
getEvents(experimentId, fcsFileId, populationId, format = "TSV", headerQ = T)

# Saves as an FCS file to disk:
getEvents(experimentId, fcsFileId, destination = "/path/to/output.fcs")

# Saves as a TSV file to disk:
getEvents(experimentId, fcsFileId, destination = "/path/to/output.tsv", format = "TSV", headerQ = T)

# Subsamples and gates to only contain events in the specified population:
getEvents(experimentId, fcsFileId, populationId, subsampling = list(preSubsampleN = 5000, seed = 1.5))
# }